+++ src/data/collections/array/d1/uninit/mod.rs
@@ -3,7 +3,7 @@
 //!
 //
 
-use crate::mem::{Bare, MaybeUninit, Storage};
+use crate::{Bare, MaybeUninit, Storage};
 
 mod methods;