+++ src/error/unwrap.rs
@@ -239,8 +239,8 @@ pub use unwrap;
 #[cfg(test)]
 mod tests {
     #[cfg(feature = "std")]
-    use crate::error::panic_catch;
-    use crate::error::{serr, sok, unwrap, OptRes};
+    use crate::panic_catch;
+    use crate::{serr, sok, unwrap, OptRes};
 
     const OPTION_SOME: Option<bool> = Some(true);
     const OPTION_NONE: Option<bool> = None;